+++ src/models/triggers/EntrySaved.php
@@ -4,9 +4,9 @@ namespace Ryssbowh\CraftTriggers\models\triggers;
 
 use Ryssbowh\CraftTriggers\Triggers;
 use Ryssbowh\CraftTriggers\models\Trigger;
-use craft\base\Element;
 use craft\elements\Entry;
-use craft\events\ModelEvent;
+use craft\events\ElementEvent;
+use craft\services\Elements;
 use yii\base\Event;
 
 class EntrySaved extends Trigger
@@ -33,10 +33,13 @@ class EntrySaved extends Trigger
     public function initialize()
     {
         $_this = $this;
-        Event::on(Entry::class, Element::EVENT_AFTER_SAVE, function (ModelEvent $e) use ($_this) {
+        Event::on(Elements::class, Elements::EVENT_AFTER_SAVE_ELEMENT, function (ElementEvent $e) use ($_this) {
+            if (!$e->element instanceof Entry) {
+                return;
+            }
             Triggers::$plugin->triggers->onTriggerTriggered($_this, [
-                'entry' => $e->sender,
-                'isNew' => $e->sender->firstSave,
+                'entry' => $e->element,
+                'isNew' => $e->isNew,
                 'event' => $e
             ]);
         });